[android] Extracted common logic for subscription flow in abstract fragment for futher reuse
